backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.4.0
python-dotenv==1.0.1
pydantic==2.6.1
dnspython==2.7.0
pymongo==4.8.0
